requirement.txt
fastapi
uvicorn
opencv-python
torch
torchvision
pillow
imagehash
numpy